Consolidate cached boxed values between Expressions and Interpreter
A utility class in Expressions has some cached instances of ConstantExpression
for common values.

A utility class in Interpreter has some cached boxed instances of the same values.

Consolidate these.

In updating `ScriptingRuntimeHelpers.BooleanToObject` accordingly I noticed that
it is always used in a way that prevents an overload that does the exact same
logic being used in InterpretedFrame.Push or InstructionList.EmitLoad (indeed the
latter skips a few steps if given a boolean directly), so I just removed that
method.