Special-case known schemes in CheckSchemeSyntax
The 32-bit special-casing is already handled by CheckKnownSchemes.  On 64-bit, without this we end up allocating a string for the scheme and doing a dictionary lookup; instead, we can special-case the common schemes to avoid that allocation and lookup.

(With a small tweak to the existing tests, this new code has 100% line and branch coverage.)