Rewrite Uri.CheckForUnicode to improve throughput and reduce allocations
Every call to this is allocating a new char[], unencoding into that array, and then iterating through the array looking for non-ASCII chars.  We can simplify all of that to avoid needing to store the results into the destination.  This avoids the allocation, the copy, and a second iteration through the results.  It also lets us bail early if we do fine unicode characters.  All tested scenarios get significantly faster.