Test Debug Proxies of Expression objects.
For various types of expression, obtain the DebuggerTypeProxy through
similar means as that used by a debugger (examine the
DebuggerTypeProxyAttribute and instantiate the type), then compare
the properties obtainable through the proxy with the properties of
the expression object itself.

Failure to find a debug property for any public property is
considered a failure, with the explicit exceptions of ArgumentCount
and ExpressionCount.