Reduce allocations in HttpClient.GetString/ByteArrayAsync
Today, HttpClient.GetStringAsync and GetByteArrayAsync copy all of the response data to a MemoryStream.  For strings, that buffer is never exposed, so it ends up being temporary garbage.  For byte arrays, the data is extracted from the MemoryStream with ToArray, which means double the allocations.

This commit does three things:
    - It introduces a helper Stream based on ArrayPool, where all buffers come from the shared pool.
    - GetStringAsync uses that stream as the target, such that the byte[]s used are recycled with the pool, avoiding per-request allocations.
    - GetByteArrayAsync uses that stream as the target if a content length couldn't be determined, so that the byte[]s incurred as the stream grows are recycled.  And if a content length can be determined, it uses a MemoryStream and then returns that MemoryStream's buffer directly to the caller, avoiding the extra buffer allocation and copy.