Remove exception-type-examination kludge from S.L.E test
Previously a test used a combination of ThrowsAny and testing
ex.GetType().FullName as a kludge because referencing
RuntimeWrappedException wouldn't work.

This difficulty has been removed, so change the test to the
standard XUnit pattern for testing for a particular type of
exception.

Also test the WrappedException is correct. Which would have been
more cumbersome to do, before.