Rewrite ConcurrentQueue for better performance
This commit rewrites ConcurrentQueue<T> to provide better throughput and a much better allocation profile.  As with the previous implementation, it's structured as a linked list of segments.  Unlike the previous implementation, these segments are now ring buffers, allowing them to be reused as long as they're not full; if an enqueue does make a segment full, then a new segment is created for subsequent enqueues.  Also unlike the previous implementation, segments are dynamic in size, so newly allocated segments follow a doubling scheme similar to that employed by other generic collections.  The synchronization mechanism used is also lower-overhead than before, with a typical operation incurring just one CAS operation to either enqueue or dequeue.