Fix ConcurrentBag.ToArray ordering
Technically ConcurrentBag is unordered and thus the order of elements from its ToArray/GetEnumerator/CopyTo/etc. shouldn't matter.  But regardless I inadvertently changed the order as part of some recent reimplementation, and then added some tests which relied on the new ordering, which then break on desktop.  It's simple enough to use the same ordering and fix the tests accordingly.