Reduces branching in VisitAndConvert and similar.
Each of these methods continuously tests for the existence of an
array to see whether it should add to it. Failing that it tests to see
whether it should create it.

Instead just test as to whether it should create it, and then finish
the loop expecting it to exist.