Fix DateTimeFormatInfo test
GetInstance test depends on DateTimeFormatInfo.CurrentInfo which depends on CurrentCulture. it is not guarantees calling CurrentInfo twice to return the exact same instance because 2 threads can race in initializing current culture and we can end up overriding the current culture which cause having 2 CurrentInfo objects.
The fix here is to test the values if we are getting 2 different instances