Fix race condition in ConcurrentQueue.TryPeek
We incorrectly loop through the segments looking for the first that's not empty, under the idea that if there's more than one segment, all segments before the last are frozen, thus if we go through each segment in turn we can return from the first that's not empty. The problem is that we could check segment s, see that it's empty, but before we're able to check its next segment, so many enqueues could come in that they not only fill this segment, they spill into the next, causing the s._nextSegment to be non-null, at which point our loop will continue and we'll peek from s._nextSegment instead of from s.