Fix ConcurrentBag enumeration undefined behavior
ConcurrentBag.GetEnumerator used to copy all of the contents into a List and then return that List's enumerator.  As part of some recent optimizations, I changed that to use an array instead of a list.  But array's enumerator has different semantics for the undefined case of accessing Current after MoveNext returns false.  Initially I thought we'd just accept the breaking change, but as it's simple to address and as the change does introduce an exception where there wasn't previously one, I've changed my mind and am just adding a simple enumerator to address the discrepancy.