Initial set of work for building external dependencies
This adds an external folder with some basic infrastructure to
template project.json files by TFM and RID so that we can only
restore the set that we actually need.

Updates our configuration generation to add RuntimeOS and ArchGroup
so we can use those to build up a NugetRuntimeIdentifier for using
to restore the builds.