Clean-up batch restoring of packages.
This cleans out a lot of the project.json batch restoring, validation,
and generation from the build.proj. Also enables restoring of packages
for individual projects again. We well eventually move all projects that
need restoring under external so we can keep them isolated from the rest
of the build.