Make the Crypto.Encoding library work on macOS without OpenSSL
Apple's Security framework doesn't seem to expose any OID resolution or
ASN.1/DER pretty-printing, so the "make a native call here" functions just
always report that they have failed.

The pretty-printing of SubjectAlternativeName was manually written into this
library to retain support for libraries which are parsing the ToString output.