Update documentation to include how to configure builds
Updates the documentation as well as config.json file with information on
RuntimeOS and the different ways to configure BuildConfiguration.

Cleans up InputOSGroup as it is no longer needed.

Cleans up TargetOS and makes it scoped only to the test command line generation

Add debugging information for TargetOS property for test runs.