Add StateChangeHasSubscribers to DbConnection
ADO.NET providers are supposed to call DbConnection.OnStateChange on
connection state changes. The problem is that this design forces the
allocation of the StateChangeEventArgs parameter regardless of whether
there are subscribers to StateChange or not.

Added a simple getter to check whether subscribers exist or not. If
not calling OnStateChange can be skipped altogether.