Serialize all properties whether they are primitives or not
The purpose of this change is ot make Discovering the information that DiangosticSource
exposes to EventSources more discoverable.

Previously when sending DiagnositicSource information to EventSource targets if the
property in the payload had a type that was not a primitive or string, it was skipped
(since it was unclear how to serialize it).      The problem with this is that any unserializable
values become 'invisible', and you have to consult the source code to determine what
is available.

We remove the 'string or primitive' condition, and simply uniformly use the object's
ToString() method to serialize it.    For strings and primitives this already does what
users expect, and for 'unserializable' objects typcially the 'ToString() will be inherited
from Object, which will print the type name.    This is useful because it now shows
EventSource clients that the property exists, and even what type it is, so they know
what fields are available for creating transform specifications.