Default configuration select package rid (#15037)
* Default Configuration to BuildConfiguration

When a project does not define BuildConfigurations, its Configuration
should be set to BuildConfiguration.

Previously we weren't setting Configuration at all in this case, so we
would get default values for all the components when we tried to parse
out the Configuration string.

* Remove workaround for building pkgprojs

Now that we always set Configuration this is not needed.

* Derive package RID from RuntimOS

We have a few cases where we want to package as a different RID
than what we are running on.  I previously only special cased one
of them (win7).  This makes the code more generic and handles
everything.