Refactor build traversal
Change contract of cross-project target to return configuration rather
than trying to construct a project item.  This preserves the metadata of
the original Project/ProjectReference.  When the MSBuild task creates
the TargetOutputs it will copy all the original metadata from the
Project items and create the OriginalItemSpec metadata.  From this
we can simply transform back to a Project/ProjectReference and prepend
Configuration to the AdditionalProperties.  This fixes a case where
projects were setting AdditionalProperties and we were squashing it.

This exposed an issue in traversal where we were still setting TestTFM
(previously this was squashed, only with other AdditionalProperties).

To make it clearer how this worked I made the targets only return a
property so that we don't accidentally squash any metadata set on the
project.

I've also cleaned up the targets to make things a bit more consistent
and added comments.