Add PAL errno values for ENOTRECOVERABLE and EOWNERDEAD
These values are defined in posix. They were removed from corefx
in 61254e6e4f227b93f44a6df08270419f8d4dc36a because they are not
available on NetBSD.

This change adds them back to the PAL layer and only map them to
the native value when the platform supports it.