Make X509Extensions OID lookup work on macOS
The built-in derived types of X509Extensions use an internal constructor which
makes use of the OidGroup lookup on Oid.FromOidValue. But Oid.FromOidValue
throws when it has no match, and so the X509Extensions derived types fail.

Now the four special extensions we support will correctly resolve OID names,
using the static value of what is currently defined for OpenSSL for those OIDs.