Add RuntimeOS packages - Private.CoreFx.NETCoreApp
These RIDs were added to the native shims between when we branched
dev/eng and merged back to master.  We don't have an official build leg
building these, but I'm adding them to the identity package so that our
officially built identity package will have entries for them in its
runtime.json.  This is on-par with how the native shims were behaving.