Use BuildConfigurations for external
This modifies our external projects to use the configuration system.

To make this work I had to remove a stale condition in
buildvertical.targets.

I also noticed that unlike `netcoreapp` which mapped to `netstandard`,
`netfx` was mapping to `nestandard2.0`.  This meant that when building
for `netfx`, `netstandard` configurations would be ignored, which is
wrong.