Ensure corelib-based facades are in NETCoreApp
Anything with type-forwards into corlib needs to be part of the shared
framework.  It should also not ship as a package since it could get out
of sync with the corlib, so I've removed these packages as well and
ensured that they are part of the NETCoreApp targeting pack (removing
IsNETCoreAppRef=false).

There is one exception which is System.Buffers.  In this case it will
also ship as a package in order to provide downlevel support.  The
version in NETCoreApp will have precedence over the pacakge version
and we should consider omitting the NETCoreApp build from the package
once we bring that back.