Change tar path for Ubuntu Runs
In the recent dev/eng work we changed around the layout of the binaries
that we dropped.  But we also produce the new runtime folder which has
all of the dependencies in one place.  For Ubuntu performance runs on
coreclr we have used the uploaded zip file as our CoreFX.  This change
simply removes all of the folders that were previously packaged and now
just includes the content of the runtime folder.