Better handle synchronously completed request content ReadAsStreamAsync
Regardless of whether we already had a content stream, we were allocating a closure object on every call to the transfer function.

And for a common case of ReadAsStreamAsync completing synchronously (as it does with StreamContent), we were unnecessarily allocating several tasks as well as a delegate.  We can instead add a fast path that does the logic synchronously when the task completes synchronously.