Make ReferenceFromRuntime use project reference
ReferenceFromRuntime for src projects isn't safe.  It allows any source
project to reference the output of another source project without
properly sequencing the build.

This fixes that by changing ReferenceFromRuntime to instead use a
ProjectReference to the runtime project for the building configuration
and filter that to just the files requested.

I also completely block the use of ReferenceFromRuntime in test projects