Change UpdateVSConfiguration tasks to add ProjectGuild and empty configs
In order to generate stable solution files the projects need guids otherwise
VS will keep adding them or not be able to reuse the same guild each time we
generate the solutions. So this update task adds a guid to the project files.

For projects that don't have a configuration.props file we always add a default
Debug/Release configuration to make projects that are loaded in our solutions at
least have that consistent set of configurations.