Move perf runner to dotnet core
This check-in along with the updates to PerfTesting.targets in the build
tools repo gives us the ability to run performance tests, aggreagate the
data and submit it to Benchview.  This commit also adds the source and
project.json of the performance runner that works on dotnet core instead
of running against the Desktop framework.