Optimise S.L.Expressions compilation of primitive constants.
Most types use conv instructions but when the value is known to be in
the types range (as these are ipso facto being typed as such) this has
no effect. Just use EmitInt for all values < 32bit size.

Since EmitInt is now a bit of a misnomer, rename all such methods to
EmitPrimitive.

Use shorter ldc* instructions for 64-bit constants when the value will
fit within 32 bits, as either a signed or unsigned value.