Update Reference and ProjectReference ordering in our targets.
Switch the to using a common DefaultReference notion between targets
and eliminated targetingpack.props. Moved the default references for
tests into tests.targets.

Fixed where we added Private to References because in some cases it wasn't happing
because the target that set Private=true was run before all the references were
added. This is why we are now setting Reference Private=false in a different target
from ProjectReference Private=false because they sometimes need to happen at different
times in the target execution.

Set _FindDependencies=false to block RAR from trying to automatically
add the closure of the primary references as it added automatic references that folks
could accidently start depending on and we want to be explicit about our dependencies.
This was really noticable once we started auto referencing the netstandard facade which
caused everything to get pulled into the reference set.

To fix the various build scenarios where we add ProjectReferences automatically we needed
a common target to hook on which we now have AddProjectReferencesDynamically target that
correctly hooks into the various dependency chains. See comment in targets file for more
details.