S.L.Expressions use short ldind/stind for more types.
#11407 changed EmitLoadValueIndirect and EmitStoreValueIndirect from
if-else ladders on type == to a switch on TypeCode. The original if-
else ladder skipped many types, presumably trading off the larger IL
stream caused by using ldobj and stobj for a quicker walk through the
ladder.

Since advantage doesn't hold for a switch, add in the omitted types,
and re-order by size to make the completeness clearer.