Remove offset labels from S.L.Expression test IL dumps
Instead of labelling every IL instruction with a label based on its
offset, label only those that are targets of breaks, leaves, etc. so
that changes to expected IL to match changes in implementation can be
more localised only to those directly affected.

Clean up some unused code, etc. in the IL printer.

Contributes to #16068