Fix unreliable test in SerialPortRegressions
The read event handler in this test assumes that all the data will arrive in the first read event, and so the predicate only waited on one event.  This is a false assumption: it's unknowable how much data you'll get on any particular receive event, and you need to be prepared to receive more than one.