Correctly handle dynamic lifted enum arithmetic
Microsoft.CSharp special-cases enum binary operations, but not when
lifted. In most cases the nett result is an expression that will have
the same outcome as if the expression had been special-cased to
similarly match the C# rules exactly.

E.g. the C# rules for adding an short-backed enum x and a short y would
convert both to int, add them and then cast the result to the short-
backed enum. The expression produced casts both to short, adds them and
casts to the short-backed enum. Since in S.L.Expressions the result of
short addition is a short (unlike in C# where it is an int) the final
cast means the result is the same either way.

In the case of addition and subtraction with byte- or sbyte-based enums
however this means the expression produced tries to add byte or sbyte
values, which is not valid in S.L.Expressions, and an
InvalidOperationException results.

Catch the case of lifted enum arithmetic, producing an expression with
the correct intermediary casts.

Fixes #16028.