Simplify and optimise S.L.Expressions type-test instructions.
NullableTypeEqualsInstruction exactly duplicates TypeEqualsInstruction.

Remove it, and use GetNonNullableType() on type argument so result is
the same for nullables and non-nullables.

Path using type-equal test for type-is on nullables will work with any
value type, so test for being a value type rather than being a nullable.

ConstantCheck.AnalyzeTypeIs call in CompileTypeIsExpression may have
found only a nullity test is necessary, in which case emit it instead
of full type test.