Add a HashSet<T>.CreateSetComparer overload taking an IEqualityComparer
null is treated as shorthand for Equality<T>.Default.

Consider two sets equal iff for each element in either there is an
element in the other that the passed comparer considers equal.

Take some short-cuts where available, but only if it results in the
precise behaviour described above.

Keep old behaviour in current overload.

Closes #14416

Offers a workaround to #12560