Config system should flow insignificant properties
Some properties are "insignificant" which means that they aren't
considered for compatibility decisions but should still flow to child
projects.

In a normal build traversal these will flow because
FindBestConfiguration finds all compatible configurations for the
BuildConfiguration then chooses the best of these (which include
insignificant property values) from the set of project
BuildConfigurations.  As a result FindBestConfiguration will flow the
insignificant properties from BuildConfiguration to the Configuration
that gets passed to child projects.

This is different than BuildAllConfigurations.  In that case we don't
call any task to select configurations, we just repeat the
configurations verbatim from the project's BuildConfigurations.
Since projects never specify these insignificant properties
(eg: ConfigurationGroup, ArchGroup) we weren't flowing the values and
instead were always choosing the default value.

This change makes it so that we first try to pull insignificant
properties from the Configuration string, but if we can't find them,
we'll then try to persist the value (along with creating any derived
properties) and only if they are unset will we try to set them to their
default value.