Make RemoteExecutorConsoleApp runs against the full framework
All tests using RemoteExecutorConsoleApp having it run under the dotnet host which is not correct for the netfx runs.
The fix here is to allow running RemoteExecutorConsoleApp directly against the full framework and uses the same Netfx runtime bits.
We have to copy the config file too to prevent assembly binding issues