Cut GetTypeInfo from ConcurrentDictionary and use switch over if-else
GetTypeInfo is no longer a necessary step to get IsValueType so remove
the call to it (contributes to #16305).

Replace if-else ladder on == typeof(…) with switch on GetTypeCode so up
to a dozen tests becomes at most 3.

This makes it no longer necessary to test for enum types and get
underlying type, so don't.

Add a 64-bit type to the types covered in testing.