Improve private package build perf
After including all source in private packages build perf is
significantly regressed.  This is due to the huge number of props and
items from those props.  When evaluating project references we were
revaluating all these once for each RID.  This was excessive since the
only thing we needed out of the project reference was the package ID.
Since that's coming from the same project, just list it explicitly.