Don't use DefaultValueInstruction for nullable types.
As the result of the call to Activator.CreateInstance will be null for
nullable types, just load null directly with the singleton
LoadObjectInstruction for null values.

Since the only test coverage of this instruction was with such a
nullable type, add tests for DefaultExpression with non-nullable struct
types.

DefaultValueInstruction takes a different path for reference types, but
is only used with value types (and now only non-nullable value types)
so replace the test for IsValueType with an assertion.