Fix Dns_GetHostEntryAsync_NoneIPAddress_Fail test on Unix
The APIs appear to be working fine; it's just that on Linux we're throwing an internal exception derived from SocketException that contains more details, and the test was expecting exactly a SocketException.  Fixed the test. On macOS, the OS successfully resolves IPAddress.None to broadcasthost, producing a valid GetHostEntry result, so I've left the test as platform-specific to not run on macOS, and added one for macOS.