Modify RemoteExecutorConsoleApp to use Environment.Exit
The purpose of RemoteExecutorConsoleApp is to invoke a single test method, wait for it to complete, and then exit.  If that test method ends up creating some foreground threads, though, and they don't exit, the app will remain alive even after the test has completed, potentially causing the test invoker to fail due to timeouts.  So instead of returning an exit code from Main, we use Environment.Exit to avoid blocking the process' completion while waiting for foreground threads.