Add HashAlgorithmName overloads for Rfc2898DeriveBytes ctor
This adds support for using (HMAC) SHA256, SHA384, SHA512 in PBKDF2.
(HMAC)MD5 isn't being enabled at this time.  We can re-evaluate that later,
but since this API change is about letting newer algorithms in it doesn't
seem like letting the older ones in is high-value.

This is staying as a managed implementation in this change, switching to
using native implementations can easily be done as a followup given the new
tests.