Properly start a CFRunLoop on Mono (#16952)
In Mono, all threads that are calling into indefinitely blocking native code must configure for themselves a method to interrupt them from another thread. We wrap the CFRunLoopRun() P/Invoke in an internal call in order to do that. (internally we simply save the RunLoop handle, which we can then use with CFRunLoopStop from another thread)