Rewrite FixedSizeReader.AsyncReadPacket as an async method
Rewrite the FixedSizeReader.AsyncReadPacket as an async void method.  It's still allocation-free if the delegated ReadAsync completes synchronously, but if it completes asynchronously, there's now less allocation (fewer and smaller objects); there's slightly less if a single read is required, but those same allocations are then reused across all reads, so if multiple reads are needed, there's a significant decrease. And the code is significantly easier to read.

Also changes FixedSizeReader to be a static class, avoiding several being allocated per SslStream.