Use IsSZArray in S.Linq.Expressions
Expressions had a IsVector() method that was a workaround for the
unavailability of IsSZArray. Now that we have that, use that instead.

Flip a some if(!exp){}else{} to if(exp){}else{}. Slightly more readable

Do test for non-array passed to ArrayLength() only if IsSZArray fails,
since anything that will fail the array check will fail that subsequent
IsSZArray check anyway, avoiding one check in the non-exceptional case.