Use CFStringGetCStringPtr and PtrToStringUTF8 for CFString
When the string is internally acknowledged as being a UTF-8 string
CoreFoundation will expose the raw pointer, and we can then use
PtrToStringUTF8 to marshal it into managed memory.  Otherwise we continue
with the fallback of asking CoreFoundation to convert it to a known format
blob.

This change restores the lower-allocating path that was present before a
dev/eng conflict in dev/apple_crypto.