Added HttpStatusCode.TooManyRequests (429)
This statuscode is often returned when quotas or throttling limits are being applied to an API. This is often done using an API platform such as Azure API Management or Nginx. Since API's are gaining popularity and more companies are introducing these limits, this statuscode will appear more often in real live scenarios and therefore deserves a representation in this enum.