Reduce AcceptAsync allocations by 20% for IPv4/6
Every AcceptAsync with a SocketAsyncEventArgs ends up calculating the buffer size necessary to store the address.  To do that, it's currently serializing the EndPoint into a SocketAddress, which entails allocating both a SocketAddress and the byte[] that backs it.  But for IPv4 and IPv6, the two most common families, we know the address size in advance, and can simply use those values rather than allocating to get them.  If in the future other families become more common, we could cache those dynamically after computing them once.