Reduce memory load of ConcurrentCollections test (#17205)
* Reduce memory load of ConcurrentCollections test

This test was hitting an OOM on the CI machines because it adds new items to a collection in a loop until a timeout is met. I modified the test to instead add a fixed number of items, regardless of time. This also reduces the total runtime of the Concurrent Collections tests from ~15 seconds to ~10 seconds on my local box.

* Increase Count Concurrency of test and move to Outerloop