Remove Socket._permittedRemoteAddress
Noticed this field was unnecessary.  It's a remnant of needing to do CAS permission checks, and was used to store an address against which a permission check was already done so it could be avoided subsequently if the same address needed to be checked.

Once we remove that field, the CheckCacheRemote function is no longer needed, as without the checks associated with that field, it's the same as SnapshotAndSerialize, so we can remove it and replace all call sites.

We can then also roll CallSerializeCheckDnsEndPoint into SnapshotAndSerialize, as the only other call site to it is a clone of SnapshotAndSerialize, so we can just use it.

Also looked for remaining mentions of permissions in comments and removed them.