Force shim assembly references to 0.0.0.0
The shims will end up being pulled into assembly closures where sometimes
they will have a higher version of a dependency then what is actually needed.
This causes msbuild (RAR) to output warnings which are just noise. For shims
we really don't need the refenced to match exact versions so we can always force
them to be the lowest version 0.0.0.0 and they will unify up to what is on
the platform.