Use a single field for cached Task SAEAs
Allocate a single object to hold all of our cached SAEA instances.  This way we don't pay for the fields if we're not using the Task methods, and the overhead of the extra single object per socket when we're already creating one or more SAEA objects per socket is minor.  This will allow us to be a bit more liberal in storing more instances (e.g. for Accept) without worrying about bloating Socket further.