Pass in assembly as name and directory to illink
The linker has a bug when consuming an XML file that it won't look any
files passed with -r.  Instead it applies the XML only to files in dirs
specified with -d.

Since our reference path will have a reference assembly that matches the
assembly we're processing this caused the XML file to be ignored.

To workaround this we'll pass in the directory of the input assembly and
then refer to it by name.