Disallow FileMode.Truncate in MMF.CreateFromFile
Using CreateFromFile on netfx with FileMode.Truncate always results in an ArgumentException whereas in netcoreapp we allow it. It seems like a sort of unintended consequence of the removal of the reliance on the FileSystremRights FileStream constructor. The ArgumentException from FileMode.Truncate was coming from that constructor where it asserts that FileMode.Truncate requires FileSystemRights.Write, but we were only passing FileSystemRights.WriteData.

I'm fixing netcore to have the same behavior as netfx, except that the ArgumentException message will be more helpful and it will also have a parameterName. Tests are also updated.